pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ocp-benchmark-mcp"
version = "1.0.0"
description = "OpenShift Cluster Performance Benchmark MCP Server"
authors = [{name = "OpenShift Benchmark Team"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "fastmcp>=1.12.4",
    "fastapi>=0.115.7",
    "uvicorn>=0.30.0",
    "asyncio>=3.4.1",
    "langgraph>=0.3.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "kubernetes>=30.1.0",
    "requests>=2.31.0",
    "PyYAML>=6.0.1",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "fpdf2>=2.7.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.8.0",
    "httpx>=0.27.0",
    "aiofiles>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "isort>=5.13.0",
    "flake8>=7.0.0",
]

[project.scripts]
ocp-benchmark-mcp = "ocp_benchmark_mcp_server:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["*"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88